+++ Rubeus/Commands/Asktgs.cs
@@ -13,6 +13,7 @@ namespace Rubeus.Commands
             bool ptt = false;
             string dc = "";
             string service = "";
+            Interop.KERB_ETYPE requestEnctype = Interop.KERB_ETYPE.subkey_keymaterial;
 
             if (arguments.ContainsKey("/ptt"))
             {
@@ -22,6 +23,33 @@ namespace Rubeus.Commands
             if (arguments.ContainsKey("/dc"))
             {
                 dc = arguments["/dc"];
+            }
+
+            if (arguments.ContainsKey("/enctype"))
+            {
+                string encTypeString = arguments["/enctype"].ToUpper();
+
+                if (encTypeString.Equals("RC4") || encTypeString.Equals("NTLM"))
+                {
+                    requestEnctype = Interop.KERB_ETYPE.rc4_hmac;
+                }
+                else if (encTypeString.Equals("AES128"))
+                {
+                    requestEnctype = Interop.KERB_ETYPE.aes128_cts_hmac_sha1;
+                }
+                else if (encTypeString.Equals("AES256") || encTypeString.Equals("AES"))
+                {
+                    requestEnctype = Interop.KERB_ETYPE.aes256_cts_hmac_sha1;
+                }
+                else if (encTypeString.Equals("DES"))
+                {
+                    requestEnctype = Interop.KERB_ETYPE.des_cbc_md5;
+                }
+                else
+                {
+                    Console.WriteLine("Unsupported etype : {0}", encTypeString);
+                    return;
+                }
             }
 
             if (arguments.ContainsKey("/service"))
@@ -42,14 +70,14 @@ namespace Rubeus.Commands
                 {
                     byte[] kirbiBytes = Convert.FromBase64String(kirbi64);
                     KRB_CRED kirbi = new KRB_CRED(kirbiBytes);
-                    Ask.TGS(kirbi, service, ptt, dc, true);
+                    Ask.TGS(kirbi, service, requestEnctype, ptt, dc, true);
                     return;
                 }
                 else if (File.Exists(kirbi64))
                 {
                     byte[] kirbiBytes = File.ReadAllBytes(kirbi64);
                     KRB_CRED kirbi = new KRB_CRED(kirbiBytes);
-                    Ask.TGS(kirbi, service, ptt, dc, true);
+                    Ask.TGS(kirbi, service, requestEnctype, ptt, dc, true);
                     return;
                 }
                 else